+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # processx development version
 
+* New options in `pty_options` to set the initial size of the pseudo
+  terminal.
+
 # processx 3.4.4
 
 * processx now works correctly for non-ASCII commands and arguments passed